Fix volcano sounds
